+++ backend/src/controllers/userController.js
@@ -1,14 +1,23 @@
 import UserService from "../services/userService.js";
-import { ValidationError, NotFoundError } from "../utils/errors.js";
+import { ValidError } from "../utils/errors.js";
 
 class UserController {
     static async createUser(req, res, next) {
         try {
-            const { name, email } = req.body; 
+            let { name, email } = req.body; 
+    
+            name = name.trim();
+            email = email.trim();
+    
             if (!name || !email) {
-                throw new ValidationError('Name and email are required');
+                throw new ValidError('Name and email cannot be empty.');
             }
-
+    
+            const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+            if (!emailRegex.test(email)) {
+                throw new ValidError('Invalid email format.');
+            }
+    
             const user = await UserService.createUser({ name, email });
             return res.status(201).json(user);
         } catch (e) {
@@ -18,8 +27,8 @@ class UserController {
 
     static async getAllUsers(req, res, next) {
         try {
-            const includeDeleted = req.query.includeDeleted === 'true';
-            const users = await UserService.getAllUsers(includeDeleted);
+            const withDeleted = req.query.withDeleted === 'true' || req.query.withDeleted === '1';
+            const users = await UserService.getAllUsers(withDeleted);
             res.status(200).json(users);
         } catch (e) {
             next(e);
@@ -28,8 +37,13 @@ class UserController {
 
     static async deleteUser(req, res, next) {
         try {
-            const { id } = req.params;
-            const result = await UserService.deleteUser(id); // Валидация на integer "null" и т.д.
+            const id = Number(req.params.id);
+
+            if (!Number.isInteger(id) || id <= 0) {
+                throw new ValidError("Invalid user ID. It must be a positive integer.");
+            }
+
+            const result = await UserService.deleteUser(id);
             return res.status(200).json(result);
         } catch (e) {
             next(e);
@@ -38,8 +52,13 @@ class UserController {
 
     static async restoreUser(req, res, next) {
         try {
-            const { id } = req.params;
-            const result = await EventService.restoreUser(id);
+            const id = Number(req.params.id);
+
+            if (!Number.isInteger(id) || id <= 0) {
+                throw new ValidError("Invalid user ID. It must be a positive integer.");
+            }
+
+            const result = await UserService.restoreUser(id);
             return res.status(200).json(result);
         } catch (e) {
             next(e);
